+++ docs/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 ## [Unreleased]
 
+### Changed
+
+-   SGR string has the header as a method
+
 ## [0.1.0] - 2024-05-12
 
 ### Added